backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
fastapi==0.116.2
h11==0.16.0
idna==3.10
pydantic==2.11.9
pydantic_core==2.33.2
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
httpx 
python-dotenv
